refactor!: address PR reviews
