+++ openedx/core/djangoapps/course_groups/partition_scheme.py
@@ -10,8 +10,8 @@ from lms.djangoapps.courseware.masquerade import (
     get_masquerading_user_group,
     is_masquerading_as_specific_student
 )
-
 from xmodule.partitions.partitions import NoSuchUserPartitionGroupError  # lint-amnesty, pylint: disable=wrong-import-order
+
 from .cohorts import get_cohort, get_group_info_for_cohort
 
 log = logging.getLogger(__name__)
